+++ Optimizer/Program.cs
@@ -17,8 +17,8 @@ namespace Optimizer
         /* DO NOT LEAVE THEM EMPTY */
 
         // Enter current version here
-        internal readonly static float Major = 3;
-        internal readonly static float Minor = 9;
+        internal readonly static float Major = 4;
+        internal readonly static float Minor = 0;
 
         internal static string GetCurrentVersionTostring()
         {